fix - 게시글 삭제 버그 수정

+++ backend/src/main/java/swith/backend/service/PostServiceImpl.java
@@ -65,6 +65,7 @@ public class PostServiceImpl implements PostService{
      * 게시글 삭제
      */
     @Override
+    @Transactional
     public void delete(Long id) {
 
         Post post = postRepository.findById(id).orElseThrow(() ->
@@ -72,11 +73,14 @@ public class PostServiceImpl implements PostService{
 
         checkAuthority(post, PostExceptionType.NOT_AUTHORITY_DELETE_POST);
 
-        for (Attachment attachment : post.getAttachments()) {
-            String uploadFileName = attachment.getUploadFileName();
-            String uploadFilePath = attachment.getUploadFilePath();
-            s3Service.deleteFile(uploadFilePath, uploadFileName);
+        if (!post.getAttachments().isEmpty()) {
+            for (Attachment attachment : post.getAttachments()) {
+                String uploadFileName = attachment.getUploadFileName();
+                String uploadFilePath = attachment.getUploadFilePath();
+                s3Service.deleteFile(uploadFilePath, uploadFileName);
+            }
         }
+
         postRepository.delete(post);
     }
